Fix io_lifetimes_use_std compilation.
